Update 3 ports' swap creation to work with dArkOS

+++ ports/duke3dawo/Duke3D - Alien World Order.sh
@@ -45,13 +45,23 @@ export SDL_GAMECONTROLLERCONFIG="$sdl_controllerconfig"
 # Ensure Swap space is prepared or eduke32 may crash or fail to launch
 if [[ $CFW_NAME == *"ArkOS"* ]] || [[ $CFW_NAME == *"ODROID"* ]]; then
     pm_message "Preparing Swap File, please wait..."
-    [ -f /swapfile ] && $ESUDO swapoff -v /swapfile
-    [ -f /swapfile ] && $ESUDO rm -f /swapfile
-    $ESUDO fallocate -l 384M /swapfile
-    $ESUDO chmod 600 /swapfile
-    $ESUDO mkswap /swapfile
-    $ESUDO swapon /swapfile
-    [ -f $GAMEDIR/timidity.cfg ] && $ESUDO rm -f $GAMEDIR/timidity.cfg
+    if [[ $CFW_NAME == "dArkOS" ]]; then
+        [ -e /dev/zram0 ] && $ESUDO swapoff -a
+        [ -e /dev/zram0 ] && $ESUDO zramctl --reset /dev/zram0
+        [ -e /dev/zram1 ] && $ESUDO zramctl --reset /dev/zram1
+        [ -e /dev/zram2 ] && $ESUDO zramctl --reset /dev/zram2
+        modprobe zram
+        $ESUDO zramctl --find --size 384M
+        $ESUDO mkswap /dev/zram0
+        $ESUDO swapon /dev/zram0
+    else
+        [ -f /swapfile ] && $ESUDO swapoff -v /swapfile
+        [ -f /swapfile ] && $ESUDO rm -f /swapfile
+        $ESUDO fallocate -l 384M /swapfile
+        $ESUDO chmod 600 /swapfile
+        $ESUDO mkswap /swapfile
+        $ESUDO swapon /swapfile
+    fi
 elif [[ "${CFW_NAME^^}" == "KNULLI" ]]; then
     pm_message "Preparing Swap File, please wait..."
     [ -f /media/SHARE/swapfile ] && $ESUDO swapoff -v /media/SHARE/swapfile
@@ -60,7 +70,6 @@ elif [[ "${CFW_NAME^^}" == "KNULLI" ]]; then
     $ESUDO chmod 600 /media/SHARE/swapfile
     $ESUDO mkswap /media/SHARE/swapfile
     $ESUDO swapon /media/SHARE/swapfile
-    [ -f $GAMEDIR/timidity.cfg ] && $ESUDO rm -f $GAMEDIR/timidity.cfg
 fi
 
 if [[ "${DEVICE_NAME^^}" == "X55" ]] || [[ "${DEVICE_NAME^^}" == "RG353P" ]] || [[ "${DEVICE_NAME^^}" == "RG40XX-H" ]] || [[ "${CFW_NAME^^}" == "RETRODECK" ]]; then
