+++ ports/rott/Rise of the Triad - The Hunt Begins.sh
@@ -30,12 +30,23 @@ export LD_LIBRARY_PATH="$GAMEDIR/libs.${DEVICE_ARCH}:$LD_LIBRARY_PATH"
 
 if [[ $CFW_NAME == *"ArkOS"* ]] || [[ $CFW_NAME == *"ODROID"* ]]; then
     pm_message "Preparing Swap File, please wait..."
-    [ -f /swapfile ] && $ESUDO swapoff -v /swapfile
-    [ -f /swapfile ] && $ESUDO rm -f /swapfile
-    $ESUDO fallocate -l 384M /swapfile
-    $ESUDO chmod 600 /swapfile
-    $ESUDO mkswap /swapfile
-    $ESUDO swapon /swapfile
+    if [[ $CFW_NAME == "dArkOS" ]]; then
+        [ -e /dev/zram0 ] && $ESUDO swapoff -a
+        [ -e /dev/zram0 ] && $ESUDO zramctl --reset /dev/zram0
+        [ -e /dev/zram1 ] && $ESUDO zramctl --reset /dev/zram1
+        [ -e /dev/zram2 ] && $ESUDO zramctl --reset /dev/zram2
+        modprobe zram
+        $ESUDO zramctl --find --size 384M
+        $ESUDO mkswap /dev/zram0
+        $ESUDO swapon /dev/zram0
+    else
+        [ -f /swapfile ] && $ESUDO swapoff -v /swapfile
+        [ -f /swapfile ] && $ESUDO rm -f /swapfile
+        $ESUDO fallocate -l 384M /swapfile
+        $ESUDO chmod 600 /swapfile
+        $ESUDO mkswap /swapfile
+        $ESUDO swapon /swapfile
+    fi
     [ -f $GAMEDIR/timidity.cfg ] && $ESUDO rm -f $GAMEDIR/timidity.cfg
 elif [[ "${CFW_NAME^^}" == "KNULLI" ]]; then
     pm_message "Preparing Swap File, please wait..."
